iPhone survey works -- minus "other" responses
